devex-1115 - rename is-migration-complete to migrationWaitHelper; add wait-for-migration lambda
