+++ shared/admin-tools/circleci/interact-with-pending-job.js
@@ -31,6 +31,7 @@ exports.approvePendingJob = async ({ apiToken, workflowId }) => {
   await axios.post(approveJob.url, {}, approveJob);
 };
 
+// TODO: delete this if we're sure it'll never be used
 exports.cancelWorkflow = async ({ apiToken, workflowId }) => {
   const cancelWorkflow = {
     headers: { 'Circle-Token': apiToken },
